src/main/java/com/example/didemo/controllers/PropertyInjectedController.java
package com.example.didemo.controllers;


import com.example.services.GreetingService;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.beans.factory.annotation.Qualifier;
import org.springframework.stereotype.Controller;

@Controller
public class PropertyInjectedController {

    @Autowired
    @Qualifier("greetingServiceImpl")
    //if field name maches bean name Spring injects GreetingServiceImpl
    //but here is needed qualifier because @Primary overrides this behaviour
    public GreetingService greetingServiceImpl;

    public String sayHello(){
        return greetingServiceImpl.sayGreeting();
    }

}
